+++ NEWS.md
@@ -3,15 +3,17 @@
 * Convert `fm_bary()` output to a dedicated `fm_bary` class, with elements 'index'
   (index of the containing simplex) and 'where' (matrix of barycentric weights).
   (version `0.20.0.9001`)
-* Add `fm_bary_simplex()` method to extract the simplex vertex indices for an
+* Add `fm_bary_simplex()` generic method to extract the simplex vertex indices for an
   `fm_bary` object.
   (version `0.20.0.9001`)
+* Add `fm_bary_loc()` generic method for converting `fm_bary` information to
+  Euclidean coordinates (version `0.20.0.9005`)
+* Simplify `fm_basis` object creation, and add `custom_classes` developers
+  vignette (version `0.20.0.9004`)
 * Remove long deprecated `inla.mesh` etc legacy methods; need to explicitly
   convert old objects. Retaining the `inla.mesh` etc class suffixes for now.
   (version `0.20.0.9001`)
 * Handle `NA` location inputs to `fm_basis.fm_mesh_1d()` (version `0.20.0.9002`)
-* Simplify `fm_basis` object creation, and add `custom_classes` developers
-  vignette (version `0.20.0.9004`)
 
 # fmesher 0.2.0